requirements.txt
pyspark
kafka-python
psycopg2

